server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-PyMongo
flask-cors
pymongo
python-dotenv